requirements.txt
Flask==3.0.0
flask-pymongo==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
pymongo==4.6.1
flasgger==0.9.7.1
flask-swagger-ui==4.11.1